Replace nbsp with space on code block copy
Code blocks render spaces as non-breaking spaces (nbsp) in order to
prevent wrapping over lines. When copied with the "Copy to Clipboard"
button, the nbsp was copied, which can not be pasted into some
applications, like the terminal, which require standard spaces.

This replaces any occurence of nbsp with a standard space character when
the copy to clipboard button is clicked.

Closes #778.